+++ _examples/django/mysql/authuseruserpermission.xo.go
@@ -8,9 +8,9 @@ import (
 
 // AuthUserUserPermission represents a row from 'django.auth_user_user_permissions'.
 type AuthUserUserPermission struct {
-	ID           int `json:"id"`            // id
-	UserID       int `json:"user_id"`       // user_id
-	PermissionID int `json:"permission_id"` // permission_id
+	ID           int64 `json:"id"`            // id
+	UserID       int   `json:"user_id"`       // user_id
+	PermissionID int   `json:"permission_id"` // permission_id
 	// xo fields
 	_exists, _deleted bool
 }
@@ -51,7 +51,7 @@ func (auup *AuthUserUserPermission) Insert(ctx context.Context, db DB) error {
 	if err != nil {
 		return err
 	} // set primary key
-	auup.ID = int(id)
+	auup.ID = int64(id)
 	// set exists
 	auup._exists = true
 	return nil
@@ -167,7 +167,7 @@ func AuthUserUserPermissionsByPermissionID(ctx context.Context, db DB, permissio
 // AuthUserUserPermissionByID retrieves a row from 'django.auth_user_user_permissions' as a AuthUserUserPermission.
 //
 // Generated from index 'auth_user_user_permissions_id_pkey'.
-func AuthUserUserPermissionByID(ctx context.Context, db DB, id int) (*AuthUserUserPermission, error) {
+func AuthUserUserPermissionByID(ctx context.Context, db DB, id int64) (*AuthUserUserPermission, error) {
 	// query
 	const sqlstr = `SELECT ` +
 		`id, user_id, permission_id ` +
