+++ _examples/django/mysql/bookstag.xo.go
@@ -8,7 +8,7 @@ import (
 
 // BooksTag represents a row from 'django.books_tags'.
 type BooksTag struct {
-	ID     int   `json:"id"`      // id
+	ID     int64 `json:"id"`      // id
 	BookID int64 `json:"book_id"` // book_id
 	TagID  int64 `json:"tag_id"`  // tag_id
 	// xo fields
@@ -51,7 +51,7 @@ func (bt *BooksTag) Insert(ctx context.Context, db DB) error {
 	if err != nil {
 		return err
 	} // set primary key
-	bt.ID = int(id)
+	bt.ID = int64(id)
 	// set exists
 	bt._exists = true
 	return nil
@@ -153,7 +153,7 @@ func BooksTagByBookIDTagID(ctx context.Context, db DB, bookID, tagID int64) (*Bo
 // BooksTagByID retrieves a row from 'django.books_tags' as a BooksTag.
 //
 // Generated from index 'books_tags_id_pkey'.
-func BooksTagByID(ctx context.Context, db DB, id int) (*BooksTag, error) {
+func BooksTagByID(ctx context.Context, db DB, id int64) (*BooksTag, error) {
 	// query
 	const sqlstr = `SELECT ` +
 		`id, book_id, tag_id ` +
