+++ _examples/django/postgres/authusergroup.xo.go
@@ -8,9 +8,9 @@ import (
 
 // AuthUserGroup represents a row from 'public.auth_user_groups'.
 type AuthUserGroup struct {
-	ID      int `json:"id"`       // id
-	UserID  int `json:"user_id"`  // user_id
-	GroupID int `json:"group_id"` // group_id
+	ID      int64 `json:"id"`       // id
+	UserID  int   `json:"user_id"`  // user_id
+	GroupID int   `json:"group_id"` // group_id
 	// xo fields
 	_exists, _deleted bool
 }
@@ -161,7 +161,7 @@ func AuthUserGroupsByGroupID(ctx context.Context, db DB, groupID int) ([]*AuthUs
 // AuthUserGroupByID retrieves a row from 'public.auth_user_groups' as a AuthUserGroup.
 //
 // Generated from index 'auth_user_groups_pkey'.
-func AuthUserGroupByID(ctx context.Context, db DB, id int) (*AuthUserGroup, error) {
+func AuthUserGroupByID(ctx context.Context, db DB, id int64) (*AuthUserGroup, error) {
 	// query
 	const sqlstr = `SELECT ` +
 		`id, user_id, group_id ` +
