+++ _examples/django/sqlserver/authuseruserpermission.xo.go
@@ -8,9 +8,9 @@ import (
 
 // AuthUserUserPermission represents a row from 'django.auth_user_user_permissions'.
 type AuthUserUserPermission struct {
-	ID           int `json:"id"`            // id
-	UserID       int `json:"user_id"`       // user_id
-	PermissionID int `json:"permission_id"` // permission_id
+	ID           int64 `json:"id"`            // id
+	UserID       int   `json:"user_id"`       // user_id
+	PermissionID int   `json:"permission_id"` // permission_id
 	// xo fields
 	_exists, _deleted bool
 }
@@ -57,7 +57,7 @@ func (auup *AuthUserUserPermission) Insert(ctx context.Context, db DB) error {
 	if err := rows.Err(); err != nil {
 		return logerror(err)
 	} // set primary key
-	auup.ID = int(id)
+	auup.ID = int64(id)
 	// set exists
 	auup._exists = true
 	return nil
@@ -145,8 +145,8 @@ func (auup *AuthUserUserPermission) Delete(ctx context.Context, db DB) error {
 
 // AuthUserUserPermissionByID retrieves a row from 'django.auth_user_user_permissions' as a AuthUserUserPermission.
 //
-// Generated from index 'PK__auth_use__3213E83FF96CAC97'.
-func AuthUserUserPermissionByID(ctx context.Context, db DB, id int) (*AuthUserUserPermission, error) {
+// Generated from index 'PK__auth_use__3213E83FE743C00D'.
+func AuthUserUserPermissionByID(ctx context.Context, db DB, id int64) (*AuthUserUserPermission, error) {
 	// query
 	const sqlstr = `SELECT ` +
 		`id, user_id, permission_id ` +
